Código refatorado em arquivos
Criados os seguintes arquivos:
- jogador.c e jogador.h: contém funções e definições de estrutura relativas ao jogador
- montros.c e monstros.h: idem, relativas à mecânica de monstros
- mapa.c e mapa.h: idem, relativas ao desenho e leitura de mapa
- ataque_dano.c e ataque_dano.h: idem, relativas às mecânicas de ataque, dano e vidas
- variaveis_globais.h: todas as definições de constantes foram alocadas a este arquivo,
                       visando à fácil reutilização e edição.
