+++ Zinf.depend
@@ -1,11 +1,16 @@
 # depslib dependency file v1.0
-1748227686 source:c:\users\lhcma\documents\github\zinf\main.c
+1748234644 source:c:\users\lhcma\documents\github\zinf\main.c
 	"raylib.h"
 	<string.h>
 	<stdlib.h>
 	<stdio.h>
 	<math.h>
 	<time.h>
+	"jogador.h"
+	"mapa.h"
+	"monstros.h"
+	"ataque_dano.h"
+	"variaveis_globais.h"
 
 1748115983 source:c:\users\gabriel\desktop\github\zinf\main.c
 	"raylib.h"
@@ -14,4 +19,58 @@
 	<stdio.h>
 	<math.h>
 	<time.h>
+
+1748234567 c:\users\lhcma\documents\github\zinf\jogador.h
+	"variaveis_globais.h"
+
+1748234567 source:c:\users\lhcma\documents\github\zinf\jogador.h
+	"variaveis_globais.h"
+
+1748234566 source:c:\users\lhcma\documents\github\zinf\jogador.c
+	<stdio.h>
+	"raylib.h"
+	"jogador.h"
+	"variaveis_globais.h"
+
+1748233122 c:\users\lhcma\documents\github\zinf\variaveis_globais.h
+
+1748233122 source:c:\users\lhcma\documents\github\zinf\variaveis_globais.h
+
+1748234115 source:c:\users\lhcma\documents\github\zinf\mapa.c
+	<stdio.h>
+	"raylib.h"
+	"variaveis_globais.h"
+	"mapa.h"
+
+1748234089 source:c:\users\lhcma\documents\github\zinf\mapa.h
+	"variaveis_globais.h"
+
+1748235208 c:\users\lhcma\documents\github\zinf\mapa.h
+	"variaveis_globais.h"
+
+1748234931 c:\users\lhcma\documents\github\zinf\monstros.h
+	"variaveis_globais.h"
+	"jogador.h"
+
+1748234191 source:c:\users\lhcma\documents\github\zinf\monstros.c
+	<stdio.h>
+	<math.h>
+	"raylib.h"
+	"jogador.h"
+	"monstros.h"
+	"variaveis_globais.h"
+
+1748234826 source:c:\users\lhcma\documents\github\zinf\ataque_dano.c
+	<stdio.h>
+	<math.h>
+	"raylib.h"
+	"variaveis_globais.h"
+	"ataque_dano.h"
+	"jogador.h"
+	"monstros.h"
+
+1748234895 c:\users\lhcma\documents\github\zinf\ataque_dano.h
+	"variaveis_globais.h"
+	"monstros.h"
+	"jogador.h"
