+++ main.c
@@ -5,337 +5,11 @@
 #include <math.h>
 #include <time.h>
 
-#define LARGURA 1200
-#define ALTURA 800
-#define CELULAMATRIZ 50
-#define MAXMONSTROS 10
-#define MAXBALAS 3
-
-struct Player
-{
-    int posX, posY, tamanhoPersonagem, velocidadeMovimento, vidas;
-    char dirMaisRecente, armaAtual;
-    //Variável armaAtual:
-    //P = pistola;
-    //C = chicote.
-    int balas;
-    //Número de balas da pistola. Só é relevante quando o jogar está com a pistola.
-};
-
-struct Monstros
-{
-    int quantidadeMonstros, posX[MAXMONSTROS], posY[MAXMONSTROS], enterrado[MAXMONSTROS], pontoX[MAXMONSTROS], pontoY[MAXMONSTROS], timerMovimento[MAXMONSTROS], ataqueStun[MAXMONSTROS];
-    int velocidadeMovimento, raioVisao;
-    float distPlayer[MAXMONSTROS];
-};
-
-struct Balas {
-    int posX[MAXBALAS], posY[MAXBALAS], balaVel;
-    char direcao;
-};
-
-//Funcao que recebe como parametro a matriz mapa e define a posicao inicial (x,y) do jogador conforme o mapa
-void PosicionaJogadorInicialmente (char mapa[ALTURA/CELULAMATRIZ][LARGURA/CELULAMATRIZ], int *posX, int *posY)
-{
-
-    char *p;
-
-    for(p = &mapa[0][0]; p <= &mapa[(ALTURA/CELULAMATRIZ) - 1][(LARGURA/CELULAMATRIZ) - 1]; p++)
-    {
-        if (*p == 'J')
-        {
-            *posX = ((p - &mapa[0][0]) % (LARGURA/CELULAMATRIZ))*CELULAMATRIZ;
-            *posY = ((p - &mapa[0][0]) / (LARGURA/CELULAMATRIZ))*CELULAMATRIZ;
-        }
-    }
-
-}
-
-//Lê o mapa e guarda as posições dos montros nas variáveis posX e posY da estrutura Monstros, além da quantidade de monstros no mapa inteiro
-void CriaMonstros(char mapa[ALTURA/CELULAMATRIZ][LARGURA/CELULAMATRIZ], int posX[], int posY[], int enterrado[], int timerMovimento[], int pontoX[], int pontoY[], int ataqueStun[], int *quantidadeMonstros) {
-    int i, j, iMonstro = 0;
-    for(i=0;i<MAXMONSTROS;i++) {
-        posX[i] = -1;
-        posY[i] = -1;
-        enterrado[i] = 0;
-        timerMovimento[i] = 300;
-        pontoX[i] = -1;
-        pontoY[i] = -1;
-        ataqueStun[i] = 0;
-    }
-    for(i=0;i<ALTURA/CELULAMATRIZ;i++) {
-        for(j=0;j<LARGURA/CELULAMATRIZ;j++) {
-            if(mapa[i][j] == 'M') {
-                posX[iMonstro] = j*CELULAMATRIZ;
-                posY[iMonstro] = i*CELULAMATRIZ;
-                pontoX[iMonstro] = posX[iMonstro];
-                pontoY[iMonstro] = posY[iMonstro];
-                iMonstro++;
-            }
-        }
-    }
-    *quantidadeMonstros = iMonstro;
-}
-
-//Função que calcula distância entre dois pontos
-float CalculaDist(int x1, int y1, int x2, int y2) {
-    return(sqrt(pow((x1-x2),2)+pow((y1-y2),2)));
-}
-
-//Calcula a distância entre cada monstro e o player
-void DistMonstroPlayer(int posX[], int posY[], float distPlayer[], struct Player player, struct Monstros monstros) {
-    int i;
-    for(i=0;i<monstros.quantidadeMonstros;i++) {
-        distPlayer[i] = CalculaDist(posX[i],posY[i],player.posX,player.posY);
-    }
-}
-
-//Determina se o monstro deveria se mexer ou não baseado em sua distância até o player
-int MonstroDeveMover(struct Monstros monstros, struct Player player, int i) {
-    if(abs(monstros.posX[i] - player.posX) <= monstros.raioVisao && abs(monstros.posY[i] - player.posY) <= monstros.raioVisao) {
-        return 1;
-    } else {
-        return 0;
-    }
-}
-
-//Funcao que dada uma matriz mapa, desenha os obstaculos marcados como P
-void DesenhaMapa (char mapa[ALTURA/CELULAMATRIZ][LARGURA/CELULAMATRIZ])
-{
-    char *p;
-    int posX, posY;
-
-    for(p = &mapa[0][0]; p <= &mapa[(ALTURA/CELULAMATRIZ) - 1][(LARGURA/CELULAMATRIZ) - 1]; p++)
-    {
-        if (*p == 'P')
-        {
-            posX = ((p - &mapa[0][0]) % (LARGURA/CELULAMATRIZ))*CELULAMATRIZ;
-            posY = ((p - &mapa[0][0]) / (LARGURA/CELULAMATRIZ))*CELULAMATRIZ;
-            DrawRectangle(posX, posY, CELULAMATRIZ, CELULAMATRIZ, BLUE);
-        }
-        if (*p == 'V')
-        {
-            posX = ((p - &mapa[0][0]) % (LARGURA/CELULAMATRIZ))*CELULAMATRIZ;
-            posY = ((p - &mapa[0][0]) / (LARGURA/CELULAMATRIZ))*CELULAMATRIZ;
-            DrawRectangle(posX, posY, CELULAMATRIZ, CELULAMATRIZ, PINK);
-        }
-    }
-
-}
-
-//Funcao que desenha o jogador
-void DesenhaJogador (struct Player player) {
-    DrawRectangle(player.posX, player.posY, player.tamanhoPersonagem, player.tamanhoPersonagem, LIME);
-}
-
-//Funcao que desenha os monstros
-void DesenhaMonstros(struct Monstros monstros, struct Player player) {
-    int i;
-    for(i=0;i<MAXMONSTROS;i++) {
-        if(monstros.posX[i] != -1 && monstros.posY[i] != -1 && MonstroDeveMover(monstros, player, i) == 0)
-            DrawRectangle(monstros.posX[i], monstros.posY[i], CELULAMATRIZ, CELULAMATRIZ, RED);
-        if(monstros.posX[i] != -1 && monstros.posY[i] != -1 && MonstroDeveMover(monstros, player, i) == 1)
-            DrawRectangle(monstros.posX[i], monstros.posY[i], CELULAMATRIZ, CELULAMATRIZ, ORANGE);
-    }
-}
-
-//Função que desenha a quantidade de vidas do jogador
-void DesenhaVidas(struct Player player) {
-    int i;
-    if(player.vidas > 0)
-        for(i=0;i<player.vidas;i++)
-            DrawRectangle(10+i*52,10,50,50,MAROON);
-}
-
-void DesenhaBalas(struct Balas balas) {
-    int i;
-    for(i=0;i<MAXBALAS;i++)
-        if(balas.posX[i] != -1 && balas.posY[i] != -1)
-            DrawRectangle(balas.posX[i],balas.posY[i],CELULAMATRIZ,CELULAMATRIZ,YELLOW);
-}
-
-//Funcao que recebe uma direcao (U, D, L, R), uma matriz mapa e movimenta um objeto, com restricao de movimento a obstaculos (genérica), retorna 1 se houve colisão.
-int Movimenta (char direcao, char mapa[ALTURA/CELULAMATRIZ][LARGURA/CELULAMATRIZ], int *posX, int *posY, int vel)
-{
-    switch (direcao)
-    {
-    case 'U':
-        if (*posY > 0 &&
-                (mapa[(*posY - 1)/CELULAMATRIZ][(*posX + CELULAMATRIZ - 1)/CELULAMATRIZ] != 'P' &&
-                 mapa[(*posY - 1)/CELULAMATRIZ][*posX/CELULAMATRIZ] != 'P'))
-        {
-            *posY -= vel;
-            return 0;
-        } else {
-            return 1;
-        }
-
-        break;
-    case 'D':
-        if (*posY < ALTURA - CELULAMATRIZ &&
-                (mapa[(*posY + CELULAMATRIZ + 1)/CELULAMATRIZ][*posX/CELULAMATRIZ] != 'P' &&
-                 mapa[(*posY + CELULAMATRIZ + 1)/CELULAMATRIZ][(*posX + CELULAMATRIZ - 1)/CELULAMATRIZ] != 'P')
-           )
-        {
-            *posY += vel;
-            return 0;
-        } else {
-            return 1;
-        }
-
-        break;
-    case 'L':
-        if (*posX > 0 &&
-                (mapa[*posY/CELULAMATRIZ][(*posX - 1)/CELULAMATRIZ] != 'P' &&
-                 mapa[(*posY + CELULAMATRIZ - 1)/CELULAMATRIZ][(*posX - 1)/CELULAMATRIZ] != 'P')
-           )
-        {
-            *posX -= vel;
-            return 0;
-        } else {
-            return 1;
-        }
-
-        break;
-    case 'R':
-        if (*posX < LARGURA - CELULAMATRIZ &&
-                (mapa[*posY/CELULAMATRIZ][(*posX + CELULAMATRIZ + 1)/CELULAMATRIZ] != 'P' &&
-                 mapa[(*posY + CELULAMATRIZ - 1)/CELULAMATRIZ][(*posX + CELULAMATRIZ + 1)/CELULAMATRIZ] != 'P')
-           )
-        {
-            *posX += vel;
-            return 0;
-        } else {
-            return 1;
-        }
-
-        break;
-    default:
-        return 1;
-        break;
-    }
-}
-
-//Função que returna 1 caso duas teclas não opositoras estejam sendo pressionadas ao mesmo tempo, e 0 caso contrário.
-int DuasTeclas() {
-    int i = 0;
-    if(IsKeyDown(KEY_W) && IsKeyDown(KEY_A))
-        i = 1;
-    if(IsKeyDown(KEY_W) && IsKeyDown(KEY_D))
-        i = 1;
-    if(IsKeyDown(KEY_S) && IsKeyDown(KEY_A))
-        i = 1;
-    if(IsKeyDown(KEY_S) && IsKeyDown(KEY_D))
-        i = 1;
-    return i;
-}
-
-//Aplica knockback ao levar dano. Recebe um int força e tolerancia, a qual indica a que distância do centro é preciso estar para levar knockback em duas direções (genérica)
-void Knockback(int *x1, int *y1, int x2, int y2, int forca, int tolerancia, char mapa[ALTURA/CELULAMATRIZ][LARGURA/CELULAMATRIZ]) {
-    if(x2 - *x1 > tolerancia) {
-        Movimenta('L',mapa,x1,y1,forca);
-    }
-    if(x2 - *x1 < -tolerancia) {
-        Movimenta('R',mapa,x1,y1,forca);
-    }
-    if(y2 - *y1 > tolerancia) {
-        Movimenta('U',mapa,x1,y1,forca);
-    }
-    if(y2 - *y1 < -tolerancia) {
-        Movimenta('D',mapa,x1,y1,forca);
-    }
-}
-
-//Função que detecta colisão entre player e o item de vida extra, e deleta o item em caso positivo.
-int ColisaoVida(char mapa[ALTURA/CELULAMATRIZ][LARGURA/CELULAMATRIZ], struct Player player) {
-    if(mapa[player.posY/CELULAMATRIZ][player.posX/CELULAMATRIZ] == 'V') {
-        mapa[player.posY/CELULAMATRIZ][player.posX/CELULAMATRIZ] = '-';
-        return 1;
-    } else {
-        return 0;
-    }
-}
-
-//Função que atira a bala da pistola, onde o player está olhando.
-void Atirar(struct Player player, struct Balas balas, char mapa[ALTURA/CELULAMATRIZ][LARGURA/CELULAMATRIZ], int posX[], int posY[], char *dir) {
-    int i = 0;
-    if(player.dirMaisRecente == 'U' && mapa[(player.posY - 1)/CELULAMATRIZ][player.posX/CELULAMATRIZ] != 'P') {
-        while (i < 3) {
-            if(posX[i] == -1 || posY[i] == -1) {
-                posX[i] = player.posX;
-                posY[i] = player.posY - CELULAMATRIZ;
-                *dir = 'U';
-                break;
-            }
-            i++;
-        }
-    }
-    if(player.dirMaisRecente == 'D' && mapa[(player.posY + 1)/CELULAMATRIZ][player.posX/CELULAMATRIZ] != 'P') {
-        while (i < 3) {
-            if(posX[i] == -1 || posY[i] == -1) {
-                posX[i] = player.posX;
-                posY[i] = player.posY + CELULAMATRIZ;
-                *dir = 'D';
-                break;
-            }
-            i++;
-        }
-    }
-    if(player.dirMaisRecente == 'L' && mapa[player.posY/CELULAMATRIZ][(player.posX - 1)/CELULAMATRIZ] != 'P') {
-        while (i < 3) {
-            if(posX[i] == -1 || posY[i] == -1) {
-                posX[i] = player.posX - CELULAMATRIZ;
-                posY[i] = player.posY;
-                *dir = 'L';
-                break;
-            }
-            i++;
-        }
-    }
-    if(player.dirMaisRecente == 'R' && mapa[player.posY/CELULAMATRIZ][(player.posX + 1)/CELULAMATRIZ] != 'P') {
-        while (i < 3) {
-            if(posX[i] == -1 || posY[i] == -1) {
-                posX[i] = player.posX + CELULAMATRIZ;
-                posY[i] = player.posY;
-                *dir = 'R';
-                break;
-            }
-            i++;
-        }
-    }
-}
-
-//Função que seta todas as posições de balas para 0, e a velocidade para 10
-void InicializaBalas(int posX[], int posY[], int *vel) {
-    int i;
-    for(i=0;i<MAXBALAS;i++) {
-        posX[i] = -1;
-        posY[i] = -1;
-    }
-    *vel = 10;
-}
-
-//Função que atualiza a posição das balas dependendo na orientação delas, e checa colisão com monstros/paredes, destruindo a bala e matando os monstros.
-void AtualizaBalas(struct Balas balas, struct Monstros monstros, int posX[], int posY[], int monstrosPosX[], int monstrosPosY[], char mapa[ALTURA/CELULAMATRIZ][LARGURA/CELULAMATRIZ]) {
-    int i, j;
-    for(i=0;i<MAXBALAS;i++) {
-        if(posX[i] != -1 && posY[i] != -1) {
-            Movimenta(balas.direcao, mapa, &posX[i], &posY[i], balas.balaVel);
-            if(Movimenta(balas.direcao, mapa, &posX[i], &posY[i], balas.balaVel)) {
-                posX[i] = -1;
-                posY[i] = -1;
-            }
-            for(j=0;j<MAXMONSTROS;j++) {
-                if(abs(monstrosPosX[j] - posX[i]) < 25 && abs(monstrosPosY[j] - posY[i]) < 25) {
-                    monstrosPosX[j] = -1;
-                    monstrosPosY[j] = -1;
-                    posX[i] = -1;
-                    posY[i] = -1;
-                }
-            }
-        }
-    }
-}
+#include "jogador.h"
+#include "mapa.h"
+#include "monstros.h"
+#include "ataque_dano.h"
+#include "variaveis_globais.h"
 
 int main()
 {
