+++ mapa.c
@@ -1,0 +1,30 @@
+#include <stdio.h>
+
+#include "raylib.h"
+#include "variaveis_globais.h"
+#include "mapa.h"
+
+
+//Funcao que dada uma matriz mapa, desenha os obstaculos marcados como P
+void DesenhaMapa (char mapa[ALTURA/CELULAMATRIZ][LARGURA/CELULAMATRIZ])
+{
+    char *p;
+    int posX, posY;
+
+    for(p = &mapa[0][0]; p <= &mapa[(ALTURA/CELULAMATRIZ) - 1][(LARGURA/CELULAMATRIZ) - 1]; p++)
+    {
+        if (*p == 'P')
+        {
+            posX = ((p - &mapa[0][0]) % (LARGURA/CELULAMATRIZ))*CELULAMATRIZ;
+            posY = ((p - &mapa[0][0]) / (LARGURA/CELULAMATRIZ))*CELULAMATRIZ;
+            DrawRectangle(posX, posY, CELULAMATRIZ, CELULAMATRIZ, BLUE);
+        }
+        if (*p == 'V')
+        {
+            posX = ((p - &mapa[0][0]) % (LARGURA/CELULAMATRIZ))*CELULAMATRIZ;
+            posY = ((p - &mapa[0][0]) / (LARGURA/CELULAMATRIZ))*CELULAMATRIZ;
+            DrawRectangle(posX, posY, CELULAMATRIZ, CELULAMATRIZ, PINK);
+        }
+    }
+
+}
