+++ mapa.h
@@ -1,0 +1,9 @@
+#ifndef MAPA_H_INCLUDED
+#define MAPA_H_INCLUDED
+
+#include "variaveis_globais.h"
+
+void DesenhaMapa (char mapa[ALTURA/CELULAMATRIZ][LARGURA/CELULAMATRIZ]);
+
+
+#endif // MAPA_H_INCLUDED
