+++ variaveis_globais.h
@@ -1,0 +1,10 @@
+#ifndef VARIAVEIS_GLOBAIS_H_INCLUDED
+#define VARIAVEIS_GLOBAIS_H_INCLUDED
+
+#define LARGURA 1200
+#define ALTURA 800
+#define CELULAMATRIZ 50
+#define MAXMONSTROS 10
+#define MAXBALAS 3
+
+#endif // VARIAVEIS_GLOBAIS_H_INCLUDED
